requirements.txt
crewai
openai
pyyaml

